Finished first revision PlayerStats page
Added legend data (kills, damage, wins) to the legend rows in the PlayerStats page. This is still a first revision and needs some fixing. It currently does not use a ListBox, which it should in the future. Missing data from the API (such as damage and kills) should be handled correctly and not just display nothing.