Reworked PlayerStats Page
Reworked Legends display to use ListViews. Instead of trying to display kills, wins and damage for each legend (which some legends do not have any data for), the ListView now displays all available statistics fetched from the API.

Added a converter to convert the percentile of a stat to a string that displays "Top x%" or "Bottom x%" if the percentile is above or below 50% respectively.

Reworked PlayerStatsVM to include each legend that has at least one stat available, instead of three or more.

Needs fixing: the ListView displaying the legend stats is currently not entirely responsive, this should be fixed.