Reworked PlayerStats Page
Reworked Account Stats on the PlayerStats page. Instead of all the account info being at the top of the window, I moved the account info to the left side and made the elements better match with the legend stats elements.

Todo: add two fields to do a new search and a button to return to the overview page.