Finished PlayerStats Page
Reworked player search on the PlayerStats Page to easily allow switching between API data and local mock data.

When the user presses the "FETCH PLAYER DATA" button instead of the "LOAD LOCAL DATA" button when no username and or platform is inputted, the local mock data will be loaded instead.